requirements.txt
fastapi==0.65.1
SQLAlchemy==1.4.15
SQLAlchemy-Utils==0.37.3
psycopg2-binary==2.8.6
uvicorn==0.13.4
pydantic==1.8.2
starlette==0.14.2
passlib==1.7.4
bcrypt==3.2.0
python-jose==3.2.0
pytest==6.2.4
requests==2.25.1
alembic==1.6.2
python-dotenv==0.17.1
fastapi-versioning==0.9.0
aiohttp==3.7.4
python-multipart # Docker thing